+++ extension/src/buildFileWatcher/BuildFileWatcher.ts
@@ -1,12 +1,9 @@
 import * as vscode from 'vscode';
-import { logger } from './logger';
-import { GradleTaskManager } from './taskManager';
-import { GradleTaskProvider } from './tasks/GradleTaskProvider';
-import { COMMAND_REFRESH } from './commands';
+import { logger } from '../logger';
 
 type handler = () => void;
 
-export default class BuildFileWatcher implements vscode.Disposable {
+export class BuildFileWatcher implements vscode.Disposable {
   private fileSystemWatcher?: vscode.FileSystemWatcher;
   private buildFileGlob = '**/*.{gradle,gradle.kts}';
   private handlers: handler[] = [];
@@ -42,24 +39,4 @@ export default class BuildFileWatcher implements vscode.Disposable {
     this.fileSystemWatcher?.dispose();
     this.fileSystemWatcher = undefined;
   }
-}
-
-export function registerBuildFileWatcher(
-  context: vscode.ExtensionContext,
-  taskProvider: GradleTaskProvider,
-  taskManager: GradleTaskManager
-): BuildFileWatcher {
-  const buildFileWatcher = new BuildFileWatcher();
-  buildFileWatcher.addHandler(() => {
-    vscode.commands.executeCommand(COMMAND_REFRESH);
-  });
-
-  taskProvider.onDidRefreshStart(() => buildFileWatcher.stop());
-  taskProvider.onDidRefreshStop(() => buildFileWatcher.start());
-
-  taskManager.onDidEndAllTasks(() => buildFileWatcher.start());
-  taskManager.onDidStartTask(() => buildFileWatcher.stop());
-
-  context.subscriptions.push(buildFileWatcher);
-  return buildFileWatcher;
 }
