+++ extension/src/logger/index.ts
@@ -1,0 +1,3 @@
+import { Logger } from './Logger';
+
+export const logger = new Logger();
